test: split out access points tests
